Handle refunds on NPC shop failures
Npc: when selling items for non-gold currency, clone removed items for potential refund and return them to the player if delivering the reward items fails; update autobank flow to send bank balance and improve player message. Game: validate bank debit by checking expected vs actual balance and log an error on mismatch (use debug for successful debit). Items: use unsigned bit-shifts (1u << n) for cylinder flags and define FLAG_DROPONMAP. Lua: read flags as uint32_t in luaPlayerAddMoney. Tests: prevent overwriting default coin ItemType definitions by stashing/restoring original coin types and add helper functions and includes for the test adjustments.

+++ src/lua/functions/creatures/player/player_functions.cpp
@@ -2475,7 +2475,7 @@ int PlayerFunctions::luaPlayerAddMoney(lua_State* L) {
 	const uint64_t money = Lua::getNumber<uint64_t>(L, 2);
 	const auto &player = Lua::getUserdataShared<Player>(L, 1, "Player");
 	if (player) {
-		const auto flags = static_cast<uint32_t>(Lua::getNumber<int32_t>(L, 3, 0));
+		const auto flags = Lua::getNumber<uint32_t>(L, 3, 0);
 		auto [addedMoney, returnValue] = g_game().addMoney(player, money, flags);
 		Lua::pushBoolean(L, addedMoney == money);
 		lua_pushinteger(L, addedMoney);
